NeuralLyapunov

Documentation for NeuralLyapunov.

NeuralLyapunov.LyapunovDecreaseCondition

NeuralLyapunov.LyapunovMinimizationCondition

NeuralLyapunov.NeuralLyapunovStructure

NeuralLyapunov.AsymptoticDecrease

NeuralLyapunov.DontCheckDecrease

NeuralLyapunov.DontCheckNonnegativity

NeuralLyapunov.ExponentialDecrease

NeuralLyapunov.NeuralLyapunovPDESystem

NeuralLyapunov.NonnegativeNeuralLyapunov

NeuralLyapunov.NumericalNeuralLyapunovFunctions

NeuralLyapunov.NumericalNeuralLyapunovFunctions

NeuralLyapunov.PositiveSemiDefinite

NeuralLyapunov.PositiveSemiDefiniteStructure

NeuralLyapunov.StrictlyPositiveDefinite

NeuralLyapunov.UnstructuredNeuralLyapunov

NeuralLyapunov.check_decrease

NeuralLyapunov.check_fixed_point

NeuralLyapunov.check_nonnegativity

NeuralLyapunov.check_stationary_fixed_point

NeuralLyapunov.get_decrease_condition

NeuralLyapunov.get_minimization_condition

NeuralLyapunov.local_Lyapunov

NeuralLyapunov.LyapunovDecreaseCondition — Type

LyapunovDecreaseCondition(decrease, strength, checkfixedpoint)

Specifies the form of the Lyapunov conditions to be used; training will enforce decrease(V, dVdt) ≤ strength(state, fixedpoint) The inequality will be approximated by the equation relu(decrease(V, dVdt) - strength(state, fixedpoint)) = 0.0 If checkfixedpoint is false, then training assumes dVdt(fixedpoint) = 0, but if checkfixedpoint is true, then training will enforce dVdt(fixedpoint) = 0.

If the dynamics truly have a fixed point at fixedpoint and dVdt has been defined properly in terms of the dynamics, then dVdt(fixedpoint) will be 0 and there is no need to enforce dVdt(fixedpoint) = 0, so checkfixed_point defaults to false.

Examples:

Asymptotic decrease can be enforced by requiring dVdt ≤ -C |state - fixed_point|^2, which corresponds to decrease = (V, dVdt) -> dVdt strength = (x, x0) -> -C * (x - x0) ⋅ (x - x0)

Exponential decrease of rate k is proven by dVdt ≤ - k * V, so corresponds to decrease = (V, dVdt) -> dVdt + k * V strength = (x, x0) -> 0.0

NeuralLyapunov.LyapunovMinimizationCondition — Type

LyapunovMinimizationCondition

Specifies the form of the Lyapunov conditions to be used.

If checknonnegativity is true, training will attempt to enforce V(state) ≥ strength(state, fixedpoint) The inequality will be approximated by the equation relu(strength(state, fixedpoint) - V(state)) = 0.0 If checkfixedpoint is true, then training will attempt to enforce V(fixedpoint) = 0

Examples

The condition that the Lyapunov function must be minimized uniquely at the fixed point can be represented as V(fixedpoint) = 0, V(state) > 0 when state != fixedpoint. This could be enfored by V ≥ ||state - fixedpoint||^2, which would be represented, with checknonnegativity = true, by strength(state, fixedpoint) = ||state - fixedpoint||^2, paired with V(fixedpoint) = 0, which can be enforced with checkfixed_point = true

If V were structured such that it is always nonnegative, then V(fixedpoint) = 0 is all that must be enforced in training for the Lyapunov function to be uniquely minimized at fixedpoint. So, in that case, we would use checknonnegativity = false; checkfixed_point = true

In either case, relu = (t) -> max(0.0, t) exactly represents the inequality, but approximations of this function are allowed.

NeuralLyapunov.NeuralLyapunovStructure — Type

NeuralLyapunovStructure

Specifies the structure of the neural Lyapunov function and its derivative.

Allows the user to define the Lyapunov in terms of the neural network to structurally enforce Lyapunov conditions. networkdim is the dimension of the output of the neural network. V(phi::Function, state, fixedpoint) takes in the neural network, the state, and the fixedpoint, and outputs the value of the Lyapunov function at state V̇(phi::Function, Jphi::Function, f::Function, state, fixedpoint) takes in the neural network, the jacobian of the neural network, the dynamics (as a function of the state alone), the state, and the fixedpoint, and outputs the time derivative of the Lyapunov function at state.

NeuralLyapunov.AsymptoticDecrease — Method

AsymptoticDecrease(strict; check_fixed_point, C)

Constructs a LyapunovDecreaseCondition corresponding to asymptotic decrease.

If strict is false, the condition is dV/dt ≤ 0, and if strict is true, the condition is dV/dt ≤ - C | state - fixed_point |^2

The inequality is represented by a ≥ b <==> relu(b-a) = 0.0

NeuralLyapunov.DontCheckDecrease — Function

DontCheckDecrease(check_fixed_point = false)

Constructs a LyapunovDecreaseCondition which represents not checking for decrease of the Lyapunov function along system trajectories. This is appropriate in cases when the Lyapunov decrease condition has been structurally enforced.

It is still possible to check for dV/dt = 0 at fixed_point, even in this case.

NeuralLyapunov.DontCheckNonnegativity — Method

DontCheckNonnegativity(check_fixed_point)

Constructs a LyapunovMinimizationCondition which represents not checking for nonnegativity of the Lyapunov function. This is appropriate in cases where this condition has been structurally enforced.

It is still possible to check for V(fixedpoint) = 0, even in this case, for example if V is structured to be positive for state != fixedpoint, but it is not guaranteed structurally that V(fixed_point) = 0.

NeuralLyapunov.ExponentialDecrease — Method

ExponentialDecrease(k, strict; check_fixed_point, C)

Constructs a LyapunovDecreaseCondition corresponding to exponential decrease of rate k.

If strict is false, the condition is dV/dt ≤ -k * V, and if strict is true, the condition is dV/dt ≤ -k * V - C * ||state - fixed_point||^2

The inequality is represented by a ≥ b <==> relu(b-a) = 0.0

NeuralLyapunov.NeuralLyapunovPDESystem — Method

NeuralLyapunovPDESystem(dynamics, lb, ub, spec; fixed_point, ps)

Constructs a ModelingToolkit PDESystem to train a neural Lyapunov function

Returns the PDESystem and a function representing the neural network, which operates columnwise.

The neural Lyapunov function will only be trained for { x : lb .≤ x .≤ ub }. The Lyapunov function will be for the dynamical system represented by dynamics If dynamics is an ODEProblem or ODEFunction, then the corresponding ODE; if dynamics is a function, then the ODE is ẋ = dynamics(x, p, t). This ODE should not depend on t (time t=0.0 alone will be used) and should have a fixed point at x = fixed_point. The particular Lyapunov conditions to be used and structure of the neural Lyapunov function are specified through spec, which is a NeuralLyapunovSpecification.

The returned neural network function takes three inputs: the neural network structure phi, the trained network parameters, and a matrix of inputs to operate on columnwise.

If dynamics requires parameters, their values can be supplied through the Vector p, or through dynamics.p if dynamics isa ODEProblem (in which case, let the other be SciMLBase.NullParameters()). If dynamics is an ODEFunction and dynamics.paramsyms is defined, then p should have the same order.

NeuralLyapunov.NonnegativeNeuralLyapunov — Method

NonnegativeNeuralLyapunov(network_dim, δ, pos_def; grad_pos_def, grad)

Creates a NeuralLyapunovStructure where the Lyapunov function is the L2 norm of the neural network output plus a constant δ times a function pos_def.

The condition that the Lyapunov function must be minimized uniquely at the fixed point can be represented as V(fixedpoint) = 0, V(state) > 0 when state != fixedpoint. This structure ensures V(state) ≥ 0. Further, if δ > 0 and posdef(fixedpoint, fixedpoint) = 0, but posdef(state, fixedpoint) > 0 when state != fixedpoint, this ensures that V(state) > 0 when state != fixedpoint. This does not enforce V(fixedpoint) = 0, so that condition must included in the neural Lyapunov loss function.

gradposdef(state, fixedpoint) should be the gradient of posdef with respect to state at state. If gradposdef is not defined, it is evaluated using grad, which defaults to ForwardDiff.gradient.

The neural network output has dimension network_dim.

NeuralLyapunov.NumericalNeuralLyapunovFunctions — Method

NumericalNeuralLyapunovFunctions(phi, θ, network_func, V_structure, dynamics, fixed_point, grad)

Returns the Lyapunov function, its time derivative, and its gradient: V(state), V̇(state), and ∇V(state)

These functions can operate on a state vector or columnwise on a matrix of state vectors. phi is the neural network with parameters θ. network_func is an output of NeuralLyapunovPDESystem.

The Lyapunov function structure is defined by Vstructure(networkfunc, state, fixedpoint) Its gradient is calculated using grad, which defaults to ForwardDiff.gradient.

NeuralLyapunov.NumericalNeuralLyapunovFunctions — Method

NumericalNeuralLyapunovFunctions(phi, θ, network_func, structure, dynamics, fixed_point; jac, J_net)

Returns the Lyapunov function, its time derivative, and its gradient: V(state), V̇(state), and ∇V(state)

These functions can operate on a state vector or columnwise on a matrix of state vectors. phi is the neural network with parameters θ. network_func(phi, θ, state) is an output of NeuralLyapunovPDESystem, which evaluates the neural network represented phi with parameters θ at state.

The Lyapunov function structure is specified in structure, which is a NeuralLyapunovStructure. The Jacobian of the network is either specified via Jnet(phi, _θ, state) or calculated using jac, which defaults to ForwardDiff.jacobian

NeuralLyapunov.PositiveSemiDefinite — Method

PositiveSemiDefinite(check_fixed_point)

Constructs a LyapunovMinimizationCondition representing V(state) ≥ 0 If checkfixedpoint is true, then training will also attempt to enforce V(fixed_point) = 0

The inequality is represented by a ≥ b <==> relu(b-a) = 0.0

NeuralLyapunov.PositiveSemiDefiniteStructure — Method

PositiveSemiDefiniteStructure(network_dim; pos_def, non_neg, grad_pos_def, grad_non_neg, grad)

Creates a NeuralLyapunovStructure where the Lyapunov function is the product of a positive (semi-)definite function posdef which does not depend on the network and a nonnegative function nonneg which does depend the network.

The condition that the Lyapunov function must be minimized uniquely at the fixed point can be represented as V(fixedpoint) = 0, V(state) > 0 when state != fixedpoint. This structure ensures V(state) ≥ 0. Further, if posdef is 0 only at fixedpoint (and positive elsewhere) and if nonneg is strictly positive away from fixedpoint (as is the case for the default values of posdef and nonneg), then this structure ensures V(fixedpoint) = 0 and V(state) > 0 when state != fixedpoint.

gradposdef(state, fixedpoint) should be the gradient of posdef with respect to state at state. Similarly, gradnonneg(net, Jnet, state, fixedpoint) should be the gradient of nonneg(net, state, fixedpoint) with respect to state at state. If gradposdef or gradnonneg is not defined, it is evaluated using grad, which defaults to ForwardDiff.gradient.

The neural network output has dimension network_dim.

NeuralLyapunov.StrictlyPositiveDefinite — Method

StrictlyPositiveDefinite(C; check_fixed_point, relu)

Constructs a LyapunovMinimizationCondition representing V(state) ≥ C * ||state - fixedpoint||^2 If checkfixedpoint is true, then training will also attempt to enforce V(fixedpoint) = 0

The inequality is represented by a ≥ b <==> relu(b-a) = 0.0

NeuralLyapunov.UnstructuredNeuralLyapunov — Method

UnstructuredNeuralLyapunov()

Creates a NeuralLyapunovStructure where the Lyapunov function is the neural network evaluated at state. This does not structurally enforce any Lyapunov conditions.

NeuralLyapunov.check_decrease — Method

check_decrease(cond::AbstractLyapunovDecreaseCondition)

True if cond specifies training to meet the Lyapunov decrease condition, false if cond specifies no training to meet this condition.

NeuralLyapunov.check_fixed_point — Method

check_fixed_point(cond::AbstractLyapunovMinimizationCondition)

True if cond specifies training for the Lyapunov function to equal zero at the fixed point, false if cond specifies no training to meet this condition.

NeuralLyapunov.check_nonnegativity — Method

check_nonnegativity(cond::AbstractLyapunovMinimizationCondition)

True if cond specifies training to meet the Lyapunov minimization condition, false if cond specifies no training to meet this condition.

NeuralLyapunov.check_stationary_fixed_point — Method

check_stationary_fixed_point(cond::AbstractLyapunovDecreaseCondition)

True if cond specifies training for the Lyapunov function not to change at the fixed point, false if cond specifies no training to meet this condition.

NeuralLyapunov.get_decrease_condition — Method

get_decrease_condition(cond::AbstractLyapunovDecreaseCondition)

Returns a function of V, dVdt, state, and fixed_point that is equal to zero when the Lyapunov decrease condition is met and greater than zero when it is violated

NeuralLyapunov.get_minimization_condition — Method

get_minimization_condition(cond::AbstractLyapunovMinimizationCondition)

Returns a function of V, state, and fixed_point that equals zero when the Lyapunov minimization condition is met and greater than zero when it's violated

NeuralLyapunov.local_Lyapunov — Method

get_local_Lyapunov(dynamics, state_dim; fixed_point, dynamics_jac)

Uses semidefinite programming to derive a quadratic Lyapunov function for the linearization of dynamics around fixed_point. Returns (V, dV/dt, ∇V).

If dynamicsjac is nothing, the Jacobian of the dynamics is calculated using ForwardDiff. Other allowable forms are a function which takes in the state and outputs the jacobian of dynamics or an AbstractMatrix representing the Jacobian at fixedpoint. If fixed_point is not specified, it defaults to the origin.