Internals

NeuralLyapunov.phi_to_net — Function

phi_to_net(phi, θ[; idx])

Return the network as a function of state alone.

Arguments

phi: the neural network, represented as phi(x, θ) if the neural network has a single output, or a Vector of the same with one entry per neural network output.

θ: the parameters of the neural network; θ[:φ1] should be the parameters of the first neural network output (even if there is only one), θ[:φ2] the parameters of the second (if there are multiple), and so on.

idx: the neural network outputs to include in the returned function; defaults to all and only applicable when phi isa Vector.